[D] refactor - extracting daemon fetch tests to their own file.
